fix model parallelism computation
